Update to a simplified numbered grid
Rather than using the x,y coordinates where the snake should be drawn
to determine where it is, a simpler grid system of rows and columns
starting from 0 is used. When the snake or any other object such as the
food block needs to be drawn, the x,y coordinates are worked out based
on the current grid position and the size of the snake.